+++ src/app/constants.js
@@ -1,4 +1,0 @@
-/**
- * 是否为调试环境
- */
-export const DEV_MODE = process.env.NODE_ENV !== 'production';