+++ src/components/ImagePicker/index.js
@@ -1,14 +1,26 @@
 import React from 'react';
 import { ImagePicker } from 'antd-mobile';
 import compressImage from '../../utils/compressImage';
+import * as MobileDetect from '../../utils/mobileDetect';
+
 export default class ImagePickerCompress extends React.Component {
 
     static propTypes = {
         maxWidth: React.PropTypes.number,
+        onChange: React.PropTypes.func
+    }
+
+    componentDidMount() {
+        if (MobileDetect.isAndroid) {
+            let cameras = document.querySelector('input[type=\'file\']');
+            for (let i = 0; i < cameras.length; i++) {
+                cameras[i].setAttribute('capture', 'camera');
+            }
+        }
     }
 
     static defaultProps = {
-        maxWidth: 1024
+        maxWidth: 1080
     }
 
     onChange = (files, type, index) => {