+++ src/components/Navigation/index.js
@@ -1,233 +1,32 @@
 import React, { PropTypes } from 'react';
-import {
-    NavBar,
-    Icon,
-    Popover,
-    Menu,
-    Badge
-} from 'antd-mobile';
-import MobileDetect from '../../utils/mobileDetect';
-const Item = Popover.Item;
-const inlineStyles = {
-    fooMenu: {
-        zIndex: 2,
-        position: 'absolute',
-        top: 91,
-        left: 0,
-        right: 0
-    },
-    menuActive: {
-        content: '',
-        position: 'absolute',
-        width: '100%',
-        height: '100%',
-        backgroundColor: 'rgba(0,0,0,0.4)',
-        zIndex: 1
-    },
-    leftContent: {
-        width: 200,
-        height: 90,
-        display: 'flex',
-        flexDirection: 'row',
-        alignItems: 'center',
-    },
-    message: {
-        display: 'flex',
-        justifyContent: 'center',
-        alignItems: 'center',
-        width: 60,
-        height: 90,
-    },
-    menu: {
-        marginLeft: 15,
-        display: 'flex',
-        alignItems: 'center',
-        width: 60,
-        height: 90
-    }
-};
+import { NavBar } from 'antd-mobile';
+import * as MobileDetect from '../../utils/mobileDetect';
 
 export default class Navigation extends React.Component {
 
-    static PropTypes = {
-        menuData: PropTypes.array,
-        menuValue: PropTypes.string,
-        addOptions: PropTypes.array,
-        homeMode: PropTypes.bool,
+    static propTypes = {
+        onBack: PropTypes.func,
         title: PropTypes.string,
-        company: PropTypes.array,
-        companyData: PropTypes.array,
-        companyValue: PropTypes.string,
-        messages: PropTypes.array,
-        onMessagesClick: PropTypes.func
     }
 
-    constructor(props) {
-        super(props);
-        this.state = {
-            showPopover: false,
-            switchCompany: false,
-            showMenu: false,
-            navTitle: props.title || this.getNavTitle(props.menuValue, props.menuData)
-        };
+    static defaultProps = {
+        title: ' '
     }
 
-    onPopoverSelect = (opt) => {
-        this.setState({
-            showPopover: false,
-        });
-        this.props.onPopoverSelect(opt.props.value, 0);
-    }
-
-    onPopoverClick = (showPopover) => {
-        this.setState({
-            showPopover,
-            showMenu: false
-        });
-    }
-
-    getNavTitle = (value, data) => {
-        let label = '';
-        data.forEach((dataItem) => {
-            if (dataItem.value === value[0]) {
-                label = dataItem.label;
-                if (dataItem.children && value[1]) {
-                    dataItem.children.forEach((cItem) => {
-                        if (cItem.value === value[1]) {
-                            label += ` ${cItem.label}`;
-                        }
-                    });
-                }
-            }
-        });
-        return label;
-    }
-
-    onMenuSelect = (menuValue) => {
-        if (menuValue.toString() === this.props.menuValue.toString()) return;
-        this.setState({
-            showMenu: false,
-            navTitle: this.getNavTitle(menuValue, this.props.menuData)
-        });
-        this.props.onMenuSelect && this.props.onMenuSelect(menuValue);
-    }
-
-
-    onMenuClick = (e) => {
-        e.stopPropagation();
-        e.preventDefault(); // Fix event propagation on Android
-        this.setState({
-            showMenu: this.state.switchCompany ? true : !this.state.showMenu,
-            switchCompany: false
-        });
-    }
-
-    onSwitchCompanyClick = () => {
-        // Fix event propagation on Android
-        if (this.state.switchCompany) {
-            this.setState({
-                showMenu: !this.state.showMenu,
-                switchCompany: !this.state.switchCompany
-            });
-        } else {
-            this.setState({
-                showMenu: true,
-                switchCompany: true
-            });
-        }
-    }
-
-    onCompanySelect = (companyValue) => {
-        if (companyValue.toString() === this.props.companyValue.toString()) return;
-        this.setState({
-            showMenu: false,
-            switchCompany: false,
-        });
-        this.props.onCompanySelect && this.props.onCompanySelect(companyValue);
+    componentDidMount() {
+        document.title = this.props.title;
     }
 
     render() {
-        let offsetX = -10; // just for pc demo
-        if (/(iPhone|iPad|iPod|iOS|Android)/i.test(navigator.userAgent)) {
-            offsetX = -26;
-        }
 
-        const { showMenu, showPopover, switchCompany, navTitle } = this.state;
-        const { menuData, menuValue, addOptions, homeMode, title, companyData, companyValue, messages } = this.props;
-
-        const menu = (
-            <Menu
-                style={inlineStyles.fooMenu}
-                data={switchCompany ? companyData : menuData}
-                value={switchCompany ? companyValue : menuValue}
-                onChange={switchCompany ? this.onCompanySelect : this.onMenuSelect}
-                height={document.documentElement.clientHeight * 0.6}
-            />
-        );
-
-        let leftProps = {
-            onLeftClick: this.props.onBack
-        };
-        if (homeMode) {
-            leftProps = {
-                iconName: MobileDetect.isWechat ? null : 'left',
-                onLeftClick: MobileDetect.isWechat ? null : this.props.onBack,
-                leftContent: (
-                    <div style={inlineStyles.leftContent}>
-                        <div key="1" style={inlineStyles.menu} onClick={this.onMenuClick} >
-                            <Icon type={require('../../assets/menu.svg')} />
-                        </div>
-                        {messages.length > 0 ?
-                            <div key="2" style={inlineStyles.message} onClick={(e) => {
-                                e.stopPropagation();
-                                this.props.onMessagesClick && this.props.onMessagesClick();
-                            }}>
-                                <Badge text={messages.length} size="large" />
-                            </div> : null}
-                    </div>
-                )
-            };
-        }
-
+        const { title, onBack } = this.props;
         return (
-            <div >
+            MobileDetect.isWechat ?
+                null :
                 <NavBar
-                        {...leftProps}
-                        mode="light"
-                        rightContent={
-                            !homeMode ? null : [
-                                companyData[0].children.length > 1 ? <Icon key="switch" onClick={this.onSwitchCompanyClick} type={require('../../assets/switch.svg')} /> : null,
-                                <Popover key="pop" mask
-                                    visible={showPopover}
-                                    overlay={addOptions.map((option, i) => <Item
-                                        key={i}
-                                        value={option}
-                                        icon={<Icon type={require('../../assets/add.svg')} size="xs" />}
-                                    >{`新增${option}`}</Item>)}
-                                    popupAlign={{
-                                        overflow: { adjustY: 0, adjustX: 0 },
-                                        offset: [offsetX, 15],
-                                    }}
-                                    onVisibleChange={this.onPopoverClick}
-                                    onSelect={this.onPopoverSelect}>
-                                    <div style={{
-                                        height: '100%',
-                                        padding: '0 0.3rem',
-                                        marginRight: '-0.3rem',
-                                        display: 'flex',
-                                        alignItems: 'center',
-                                    }}
-                                    >
-                                        <Icon type={require('../../assets/new1.svg')} size="xs" />
-                                    </div>
-                                </Popover>
-                            ]}>
-                        {!homeMode ? title : navTitle}
-                    </NavBar>
-                    <div onClick={() => this.setState({ showMenu: false })} style={showMenu ? inlineStyles.menuActive : {}}>
-                </div>
-                {showMenu ? menu : null}
-            </div>
+                    mode="light" onLeftClick={onBack}>
+                    {title}
+                </NavBar>
         );
     }