+++ src/index.js
@@ -1,22 +1,24 @@
 import 'babel-polyfill';
-import request from './utils/request';
+import './index.css';
+import './utils/array';
 import MobileApp from './app/mobileApp';
 import bind from './app/bind';
 import { routerRedux } from 'dva/router';
-import { DEV_MODE } from './app/constants';
+import { DEV_MODE } from './constants';
 import ImageViewer from './components/ImageViewer';
 import ListView from './components/ListView';
+import Seperator from './components/Seperator';
+import Navigation from './components/Navigation';
+import ImagePicker from './components/ImagePicker';
+import View from './components/View';
+import FilterBar from './components/FilterBar';
+import request from './utils/request';
 import delay from './utils/delay';
 import wrapProps from './utils/wrapProps';
-import Seperator from './components/Seperator';
-import './index.css';
-import './utils/array';
 import backToHome from './utils/backToHome';
-import Navigation from './components/Navigation';
 import * as LocalStorage from './utils/localStorage';
 import * as SessionStorage from './utils/sessionStorage';
 import htmlFilter from './utils/htmlFilter';
-import ImagePicker from './components/ImagePicker';
 import * as MobileDetect from './utils/mobileDetect';
 import { compareObject } from './utils/object';
 export {
@@ -37,7 +39,9 @@ export {
     htmlFilter,
     ImagePicker,
     MobileDetect,
-    compareObject
+    compareObject,
+    View,
+    FilterBar
 };
 
 // npm config set registry https://registry.npm.taobao.org